requirements.txt
# Core Dependencies
streamlit==1.29.0
python-dotenv==1.0.0

# LLM Integration (only one is needed, but both available)
anthropic==0.18.1
openai==1.12.0

# NLP & Text Processing
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Data & Export
pandas==2.1.4
reportlab==4.0.8

# Utilities
python-dateutil==2.8.2
